requirements.txt
matplotlib >= 3.0.0
pandas >= 0.22.0
numpy >= 1.14.0
click == 7.0
pickleDB == 0.9.2
dialogflow == 0.5.2
protobuf >= 3.6.1
khiva >= 0.2.0